Can only update factions score is valid
