--- src/data/wars/the-end/problems/ending-array/data.ts
@@ -1,15 +1,0 @@
-import type { Problem } from "../../../../Problem";
-
-export const data: Problem<string, string> = {
-  name: "Ending array",
-  difficulty: "medium",
-  number: 55,
-  arguments: [{ name: "s", type: "string" }],
-  tags: ["arrays"],
-  testCases: [
-    {
-      input: [`"s"`],
-      output: "a",
-    },
-  ],
-};
