--- src/data/wars/the-end/problems/so-we-good/data.ts
@@ -1,15 +1,0 @@
-import type { Problem } from "../../../../Problem";
-
-export const data: Problem<string, string> = {
-  name: "So we good",
-  difficulty: "hard",
-  number: 56,
-  arguments: [{ name: "s", type: "string" }],
-  tags: ["arrays"],
-  testCases: [
-    {
-      input: [`"s"`],
-      output: "a",
-    },
-  ],
-};
