--- src/data/wars/the-end/problems/so-we-good/description.md
@@ -1,21 +1,0 @@
-# So we good
-
-Skriv en funktion **soWeGood(s)** som tar emot en array av strängar. Funktionen ska returnera en samma array fast med varje sträng med stora bokstäver
-
-## Exempel 1
-
-**_Input_**
-
-```bash
-s = "racecar"
-```
-
-**_Output_**
-
-```bash
-true
-```
-
-**_Förklaring_**
-
-En förklaring
