--- src/data/wars/the-end/war.ts
@@ -1,8 +1,0 @@
-import type { WarCreate } from "../../War";
-
-export const data: WarCreate = {
-  name: "The end",
-  number: 2,
-  startTime: new Date("9 feb 2023"),
-  endTime: new Date("13 feb 2023"),
-};
