--- src/data/wars/the-first/problems/fredde/data.ts
@@ -1,15 +1,0 @@
-import type { Problem } from "../../../../Problem";
-
-export const data: Problem<string, string> = {
-  name: "Fredde",
-  difficulty: "easy",
-  number: 57,
-  arguments: [{ name: "s", type: "string" }],
-  tags: ["arrays"],
-  testCases: [
-    {
-      input: [`"s"`],
-      output: "a",
-    },
-  ],
-};
