--- src/data/wars/the-first/problems/jasmin/data.ts
@@ -1,15 +1,0 @@
-import type { Problem } from "../../../../Problem";
-
-export const data: Problem<string, string> = {
-  name: "Jasmin",
-  difficulty: "medium",
-  number: 58,
-  arguments: [{ name: "s", type: "string" }],
-  tags: ["arrays", "basics", "strings"],
-  testCases: [
-    {
-      input: [`"s"`],
-      output: "a",
-    },
-  ],
-};
