--- src/data/wars/the-first/war.ts
@@ -1,8 +1,0 @@
-import type { WarCreate } from "../../War";
-
-export const data: WarCreate = {
-  name: "The first",
-  number: 1,
-  startTime: new Date("2 feb 2023"),
-  endTime: new Date("7 feb 2023"),
-};
