--- src/data/wars/we-go-again/problems/skojsigt/data.ts
@@ -1,15 +1,0 @@
-import type { Problem } from "../../../../Problem";
-
-export const data: Problem<string, string> = {
-  name: "Skojsigt",
-  difficulty: "medium",
-  number: 64,
-  arguments: [{ name: "s", type: "string" }],
-  tags: ["objects"],
-  testCases: [
-    {
-      input: [`"s"`],
-      output: "a",
-    },
-  ],
-};
