--- src/data/wars/we-go-again/war.ts
@@ -1,8 +1,0 @@
-import type { WarCreate } from "../../War";
-
-export const data: WarCreate = {
-  name: "We go again",
-  number: 3,
-  startTime: new Date("10 feb 2023"),
-  endTime: new Date("17 feb 2023"),
-};
